% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CroceaEnrich.R
\name{CroceaEnrich}
\alias{CroceaEnrich}
\title{Title: CroceaEnrich Function
Description: This function performs GO and KEGG enrichment analysis.}
\usage{
CroceaEnrich(file, orgdb, prefix = "CroceaEnrich_Result")
}
\arguments{
\item{file}{A data frame containing gene IDs.}

\item{prefix}{A character string used as the prefix for output files.}
}
\description{
Title: CroceaEnrich Function
Description: This function performs GO and KEGG enrichment analysis.
}
